* provide a way to define datestamp from outside of this script * back up only if the files have not been backed up before * provide ${datestamp}.csv to scripts so that we can run this whole script manually in a troubleshooting situation

+++ rpms_cron.sh
@@ -2,7 +2,12 @@
 
 export HOSTNAME=1200941-Prescient.orygen.org.au
 
-datestamp=$(date +"%d.%m.%Y")
+# provide a way to define datestamp from outside of this script
+if [ -z $datestamp ]
+then
+    datestamp=$(date +"%d.%m.%Y")
+fi
+
 INCOMING=/var/lib/prescient/RPMS_incoming
 BACKUP=/var/lib/prescient/one_day_backup
 export PATH=$HOME/miniconda3/bin/:/var/lib/prescient/utility/:$PATH
@@ -23,12 +28,17 @@ do
 done
 
 mkdir -p $BACKUP
-rm ${BACKUP}/*
-cp ${INCOMING}/*_${datestamp}.csv ${BACKUP}/
+# back up only if the files have not been backed up before
+count=`ls ${BACKUP}/*_${datestamp}.csv | wc -l`
+if [ $count -lt 76 ]
+then
+    rm ${BACKUP}/*
+    cp ${INCOMING}/*_${datestamp}.csv ${BACKUP}/
+fi
 
-rename_RPMS_vars.py $INCOMING && \
-replace_RPMS_values.py $INCOMING && \
-rpms_psychs_partition.py $INCOMING && \
+rename_RPMS_vars.py $INCOMING ${datestamp}.csv && \
+replace_RPMS_values.py $INCOMING ${datestamp}.csv && \
+rpms_psychs_partition.py $INCOMING ${datestamp}.csv && \
 convert_cbc_units.py $INCOMING && \
 rm $INCOMING/.complete_* && \
 touch $INCOMING/.complete_${datestamp}
